Update patch set 1
Patch Set 1: Code-Review-1

thats actually not what we talked about. you set a system property in init which contains the file name that is used instead of the default mixer_path.xml

here you only get the value from the property and build the path with it. the build flag is only true

Patch-set: 1
Label: Code-Review=-1